strlen size_t

+++ ft_strlen.c
@@ -1,8 +1,8 @@
 #include "libft.h"
 
-int	ft_strlen(const char *c)
+size_t	ft_strlen(const char *c)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (c[i] != '\0')
